Bug fix: correct prototype/function for VLAN PCP
Signature of set_pcp in declaration differs from definition
(uint16_t vs uint8_t)

Thanks Rob Vaterlaus for identifying it.

+++ src/include/netinet++/vlan.hh
@@ -67,7 +67,7 @@ vlan::pcp() const
 //-----------------------------------------------------------------------------
 inline
 void
-vlan::set_pcp(uint16_t pcp)
+vlan::set_pcp(uint8_t pcp)
 { tci = (tci & ~htons(PCP_MASK)) |
         htons((((uint16_t)pcp) << PCP_SHIFT) & PCP_MASK); }
 //-----------------------------------------------------------------------------
